Add type scheme erasure and make erasure total
This was non-trivial because we have to use information about the arity
of the type scheme to determine when to call normal erasure. In
particular, for partial application such as

Definition alias : Type -> Type := option.

We have to eta expand alias before we can erase it with normal type
erasure, since `option` needs to be applied before it is a type. To
carry all the proofs through we thus need more information about arities
than we had previously, as we now need to know how many arguments are
expected.

Thus, flag_of_type changes significantly. It now returns a proof of
whether the type is convertible or not to a `conv_arity` record that
contains information about the arity: the domains, the universe (as
informational content), plus a proof that the type is convertible to the
arity created from these.

As a result we can now make erasure total since we now have the ability
to erase all constant declarations. Previously we were also failing on
axiomatic type schemes, however now the AST allows a type alias to have
no body to handle this when printing instead.