Update for new MetaCoq with flag in evaluation relation
Update for the new MetaCoq version that has a flag that decides whether
matches on boxes can be done. The new relation also requires information
about inductives to be present to determine whether matches are
propositional, so because of this our erasure also needed to change a
bit. In particular, trans_env now has a straightforward translation for
each declaration type.

This should allow us to use MetaCoq's box optimization in a follow-up
change.