Only allow contracts access to their own balance
Remove the account_balance function from Chain and add instead
ctx_contract_balance to the ContractCallContext. This is is the only
thing our contracts use and it means that we match the models of our
target languages better. It also means we do not need Proper instances
anymore for contracts.